src/Category/infra/repository/category-in-memory.repositorie.spec.ts
import { Category } from "../../domain/entity/category";
import { CategoryInMemoryRepository } from "./category-in-memory.repositorie";

describe("CategoryInMemoryRepository", () => {
  let repository: CategoryInMemoryRepository;

  beforeEach(() => (repository = new CategoryInMemoryRepository()));
  it("should no filter items when filter object is null", async () => {
    const items = [new Category({ name: "test", description: "test" })];
    const filterSpy = jest.spyOn(items, "filter" as any);

    let itemsFiltered = await repository["applyFilter"](items, null);
    expect(filterSpy).not.toHaveBeenCalled();
    expect(itemsFiltered).toStrictEqual(items);
  });

  it("should filter items using filter parameter", async () => {
    const items = [
      new Category({ name: "test" }),
      new Category({ name: "TEST" }),
      new Category({ name: "fake" }),
    ];
    const filterSpy = jest.spyOn(items, "filter" as any);

    let itemsFiltered = await repository["applyFilter"](items, "TEST");
    expect(filterSpy).toHaveBeenCalledTimes(1);
    expect(itemsFiltered).toStrictEqual([items[0], items[1]]);
  });

  it("should sort by created_at when sort param is null", async () => {
    const created_at = new Date();
    const updated_at = new Date(created_at.getTime() + 10000000);
    console.log(created_at, updated_at);


    const items = [
      new Category({ name: "test", created_at: created_at, description: "test", updated_at: updated_at }),
      new Category({
        name: "TEST",
        created_at: new Date(created_at.getTime() + 1000),
        updated_at
      }),
      new Category({
        name: "fake",
        created_at: new Date(created_at.getTime() + 3000),
        updated_at
      }),
    ];

    let itemsSorted = await repository["applySort"](items, null, null);
    expect(itemsSorted).toStrictEqual([items[2], items[1], items[0]]);
  });

  it("should sort by name", async () => {
    const items = [
      new Category({ name: "cccc", }),
      new Category({ name: "bbbb", }),
      new Category({ name: "aaaa", }),
    ];

    let itemsSorted = await repository["applySort"](items, "name", "asc");
    expect(itemsSorted).toStrictEqual([items[2], items[1], items[0]]);

    itemsSorted = await repository["applySort"](items, "name", "desc");
    expect(itemsSorted).toStrictEqual([items[0], items[1], items[2]]);
  });
});